[two_dimensional_scrollables] Add fillExtent support to TreeView
Introduce a new boolean property to the TreeRow called fillExtent. If the treeNodeBuilder returns a Row, and fillExtent: true is set for that row, this outer Row will be forced to the viewport's width. Any Expanded child within this Row will then correctly expand to fill the available space within that viewport-wide Row.